// src/app/models/customer.model.ts
export class Customer {
    id?: number = 0;
    title: string = '';
    firstname: string = '';
    lastname: string = '';
    email: string = '';
    contactnumber: string = '';
    companyname: string = '';
    street: string = '';
    pincode: string = '';
    city: string = '';
    country: string = '';
}